LLM base class working
